Conflicts:
	Build/build/build.xml
	Examples/Misc/ERPDFExamples/.gitignore
	Frameworks/Ajax/MooTools/.classpath
	Frameworks/Ajax/MooTools/.gitignore
	Frameworks/Ajax/MooTools/.settings/org.eclipse.core.resources.prefs
	Frameworks/Ajax/MooTools/Sources/er/ajax/mootools/MTAjaxFileUpload.java
	Frameworks/Ajax/MooTools/Sources/er/ajax/mootools/MTAjaxModalContainer.java
	Frameworks/Ajax/MooTools/Sources/er/ajax/mootools/MTAjaxObserveField.java
	Frameworks/Ajax/MooTools/Sources/er/ajax/mootools/MTAjaxSubmitButton.java
	Frameworks/Core/ERExtensions/Sources/er/extensions/foundation/ERXStringUtilities.java
	Frameworks/Core/JavaWOExtensions/Sources/com/webobjects/woextensions/JSComponent.java
	Frameworks/Core/JavaWOExtensions/Sources/com/webobjects/woextensions/_ClassNameComparator.java

+++ Frameworks/Core/ERExtensions/Sources/er/extensions/appserver/ERXKeepAliveResponse.java
@@ -6,7 +6,6 @@ import java.io.UnsupportedEncodingException;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
-import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 
 import com.webobjects.foundation.NSForwardException;
@@ -70,7 +69,7 @@ public class ERXKeepAliveResponse extends ERXResponse {
 				}
 			}
 
-		}, 1, Integer.MAX_VALUE); // MS: turning it up to 11
+		}, 1, Long.MAX_VALUE); // MS: turning it up to 11
 	}
 
 	/**
